refactor(broker): extract CLI args and per-chain provider construction into args.rs
Moves CoreArgs (clap-derived top-level CLI), ChainArgs (per-chain
config), ChainPipeline<P> (per-chain resources passed to
Broker::start_service), and build_chain_provider (constructs the typed
provider + AnyNetwork provider + gas priority mode) out of lib.rs into
their own module.

Public types are re-exported via `pub use` so existing call-sites
(broker::CoreArgs, etc.) keep working unchanged.

lib.rs imports trimmed to drop pieces only used by the extracted code:
the alloy provider stack (RpcClient, Http, ProviderBuilder,
RetryBackoffLayer, EthereumWallet, ChainIdFiller, parse_ether),
boundless_market provider helpers (BalanceAlertConfig/Layer,
DynamicGasFiller, NonceProvider, PriorityMode), the broker-internal
RpcMetricsLayer / SequentialFallbackLayer, tower's Layer/ServiceBuilder,
clap::Parser, and Duration.

One test in prioritization.rs referenced `crate::PrivateKeySigner`
(previously in scope through lib.rs's import). It now uses the explicit
`alloy::signers::local::PrivateKeySigner` path.

Net: lib.rs shrinks by ~250 lines.
